Update d1.php
the query to return subjects and display them in the student d1 view relies on there being at least two published modules. this is not the desired functionality so have changed this to check to see at least one published module so that at least the subject i snot empty.